Added full testing for subworkers (does not run on Chrome as not supported = Test will pass but print message to console).

+++ tests/_base/configWebWorker/worker4.js
@@ -3,8 +3,33 @@
 //      messages between them.
 
 var gloabl = this;
-var worker = new Worker("worker4-1.js");
-worker.addEventListener("message", function(message) {
-    gloabl.postMessage(message.data);
-    worker.terminate();
-}, false);
+
+var dojoConfig = {
+    "baseUrl":"../../../../",
+    "async": true,
+    "packages":[
+        {"name":"dojo", "location":"dojo"}
+    ]
+}
+
+importScripts("../../../dojo.js");
+require(["dojo/has"], function(has){
+    // Test for workers, currently chrome does not support subworkers.
+
+    has.add("webworkers", (typeof Worker === 'function'));
+    if(has("webworkers")){
+        var worker = new Worker("worker4-1.js");
+        worker.addEventListener("message", function(message) {
+            gloabl.postMessage(message.data);
+            worker.terminate();
+        }, false);
+    }else{
+        this.postMessage({
+            "test":"Platform does not support subworkers",
+            "value":true,
+            "info":true
+        });
+    }
+});
+
+
